Prevent subnetwork deletion if there are IPs allocated in it
